feat: add solutions to lc problem: No.2359
No.2359.Find Closest Node to Given Two Nodes
