+++ solution/2300-2399/2359.Find Closest Node to Given Two Nodes/Solution.go
@@ -7,27 +7,27 @@ func closestMeetingNode(edges []int, node1 int, node2 int) int {
 		}
 	}
 	const inf int = 1 << 30
-	dijkstra := func(i int) []int {
+	f := func(i int) []int {
 		dist := make([]int, n)
 		for j := range dist {
 			dist[j] = inf
 		}
 		dist[i] = 0
-		q := hp{}
-		heap.Push(&q, pair{0, i})
+		q := []int{i}
 		for len(q) > 0 {
-			i := heap.Pop(&q).(pair).i
+			i = q[0]
+			q = q[1:]
 			for _, j := range g[i] {
-				if dist[j] > dist[i]+1 {
+				if dist[j] == inf {
 					dist[j] = dist[i] + 1
-					heap.Push(&q, pair{dist[j], j})
+					q = append(q, j)
 				}
 			}
 		}
 		return dist
 	}
-	d1 := dijkstra(node1)
-	d2 := dijkstra(node2)
+	d1 := f(node1)
+	d2 := f(node2)
 	ans, d := -1, inf
 	for i, a := range d1 {
 		b := d2[i]
@@ -39,12 +39,3 @@ func closestMeetingNode(edges []int, node1 int, node2 int) int {
 	}
 	return ans
 }
-
-type pair struct{ d, i int }
-type hp []pair
-
-func (h hp) Len() int           { return len(h) }
-func (h hp) Less(i, j int) bool { return h[i].d < h[j].d }
-func (h hp) Swap(i, j int)      { h[i], h[j] = h[j], h[i] }
-func (h *hp) Push(v any)        { *h = append(*h, v.(pair)) }
-func (h *hp) Pop() any          { a := *h; v := a[len(a)-1]; *h = a[:len(a)-1]; return v }
